// projects/roolith-ng/src/lib/components/badge/data-access/badge.interface.ts
export type BadgeType =
  | 'primary'
  | 'success'
  | 'danger'
  | 'warning'
  | 'info'
  | 'large'
  | 'subtle'
  | 'intense'
  | 'default';
export type BadgeSizeType = 'large' | 'default';
export type BadgeVariant = 'default' | 'status';
export type BadgeIconColor = 'primary' | 'secondary' | 'success' | 'danger' | 'warning' | 'info' | 'default';